:Refactoring
Testing the layout
