+++ app/page.js
@@ -1,0 +1,24 @@
+export default function Home() {
+  return (
+    <>
+
+<div className="navbar bg-base-100">
+  <div className="flex-1">
+    <section className="normal-case text-xl">trackdemon.in</section>
+  </div>
+  <div className="flex-none">
+  <button class="btn btn-square btn-ghost">
+      <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" class="inline-block w-5 h-5 stroke-current"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M4 6h16M4 12h16M4 18h16"></path></svg>
+    </button>
+  </div>
+</div>
+<footer class="footer items-center p-4 bg-neutral text-neutral-content">
+  <aside class="items-center grid-flow-col">
+
+  </aside>
+  </footer>
+
+</>
+
+  )
+}
