+++ styles/grid-check.css
@@ -1,0 +1,16 @@
+.page-grid {
+    width: 95%;
+    max-width: 1600px;
+    grid-column-gap: 32px;
+    grid-row-gap: 32px;
+    grid-template-rows: minmax(328px, auto) minmax(328px, auto) minmax(
+        328px,
+        auto
+      );
+    grid-template-columns: minmax(260px, 1fr) 1fr 1fr minmax(220px, 1fr);
+    grid-auto-rows: minmax(328px, max-content);
+    grid-auto-columns: 1fr;
+    margin: 32px auto;
+    display: grid;
+  }
+
